Fixed detection metrics different architecctures
Renamed detection_metrics_different_thresholds_cnn to detecctions_metrics_different_arquitectures. This file creates a .mat file with the properties of the corresponding detections (model and th).
Collect properties uses these files and creates a structure with all the ripple properties, for each model and threshold. Then a smaller and more representative properties file is created with the average and median properties of each session.